Update to xbmc/xbmc@bff0ea8a84dc5ff96c24bd45deb22e8d8fe59dba

+++ xbmc/addons/AddonCallbacksAddon.h
@@ -1,7 +1,7 @@
 #pragma once
 /*
  *      Copyright (C) 2012-2013 Team XBMC
- *      http://www.xbmc.org
+ *      http://xbmc.org
  *
  *  This Program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -47,9 +47,9 @@ public:
   // file operations
   static void* OpenFile(const void* addonData, const char* strFileName, unsigned int flags);
   static void* OpenFileForWrite(const void* addonData, const char* strFileName, bool bOverwrite);
-  static unsigned int ReadFile(const void* addonData, void* file, void* lpBuf, int64_t uiBufSize);
+  static ssize_t ReadFile(const void* addonData, void* file, void* lpBuf, size_t uiBufSize);
   static bool ReadFileString(const void* addonData, void* file, char *szLine, int iLineLength);
-  static int WriteFile(const void* addonData, void* file, const void* lpBuf, int64_t uiBufSize);
+  static ssize_t WriteFile(const void* addonData, void* file, const void* lpBuf, size_t uiBufSize);
   static void FlushFile(const void* addonData, void* file);
   static int64_t SeekFile(const void* addonData, void* file, int64_t iFilePosition, int iWhence);
   static int TruncateFile(const void* addonData, void* file, int64_t iSize);
