+++ xbmc/filesystem/TuxBoxFile.h
@@ -1,7 +1,7 @@
 #pragma once
 /*
  *      Copyright (C) 2005-2013 Team XBMC
- *      http://www.xbmc.org
+ *      http://xbmc.org
  *
  *  This Program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -34,7 +34,7 @@ namespace XFILE
       virtual void Close();
       virtual bool Exists(const CURL& url);
       virtual int Stat(const CURL& url, struct __stat64* buffer);
-      virtual unsigned int Read(void* lpBuf, int64_t uiBufSize);
+      virtual ssize_t Read(void* lpBuf, size_t uiBufSize);
       virtual int64_t Seek(int64_t iFilePosition, int iWhence = SEEK_SET);
     protected:
   };
